requirements.txt
streamlit
opencv-python
numpy
scikit-image
pandas
matplotlib
plotly
requests
python-dotenv



# pip install -r requirements.txt
# streamlit run main.py

